test: use llm SetClient in gemini tests
